+++ Changes.md
@@ -10,6 +10,7 @@ New features
 - Add `Binary` / `LargeBinary` support for `arrow2`
 - Add support to serialize / deserialize `bool` from integer arrays
 - Add a helper to construct `Bool8` arrays
+- Include the path of the field that caused an error in the error message
 
 API changes